backend/src/requirements.txt
google-cloud-pubsub
textblob==0.17.1
google-cloud-pubsub 
google-cloud-firestore